20250524 유지보수를 위한 모듈화, 위젯분리, 모델 정의, 서비스 분리
