+++ lib/utils/region_grid_map.dart
@@ -1,0 +1,11 @@
+const Map<String, Map<String, int>> regionGridMap = {
+  '내 위치': {},
+  '서울': {'nx': 60, 'ny': 127},
+  '부산': {'nx': 98, 'ny': 76},
+  '대구': {'nx': 89, 'ny': 90},
+  '인천': {'nx': 55, 'ny': 124},
+  '광주': {'nx': 58, 'ny': 74},
+  '대전': {'nx': 67, 'ny': 100},
+  '울산': {'nx': 102, 'ny': 84},
+  '세종': {'nx': 66, 'ny': 103},
+};
